Fix PathExpression type resolution
There are cases where the name resolver can fully resolve a path such as:

```rust
struct Foo(i32);

impl Foo{
  fn new() { ... }
}

fn main() {
  let a;
  a = Foo::new();
}
```

This case does not require a PathProbe. This enhances the Path resolution
to iterate each segment untill we are unable to resolve the root segment
at that point we can fall back to a PathProbe and lookup any impl item that
might satisfy the path instead.

Addresses #432